backend/fastapi_web/requirements.txt
fastapi
uvicorn
psycopg2-binary
boto3
python-dotenv
openai
anthropic
requests